add get absolute method and add WINDOWS marco ,but no ok

+++ include/lthread.h
@@ -49,7 +49,7 @@ class lthread:public Basethread
 		friend class logT;
 		//~lthread();
 		int run();
-		long getfsize(FSU& fsu);		
+		float getfsize(FSU& fsu);
 		bool checkffull();
 		bool checklogfnum(string logpath);
 };
